Skips ci when tagging a release
Tagging EE, is currently triggering a merge train to sync changes to
GitLab-FOSS, which causes repositories to be out of sync just before
tagging process is finished.

Detecting this on merge train might be a difficult task, so we opted
for adding skip ci when tagging, just like we do when adding the
changelog.

Related to https://gitlab.com/gitlab-org/release-tools/issues/358