Refactor how `ComponentVersions` deals with CNG maps
The crux of this functionality is that CNG variables are just the
Omnibus `*_VERSION` files with some extra considerations:

1. Some variables differ from the filenames (`VERSION` becomes
   `GITLAB_VERSION`)
2. Versions of GitLab components such as Gitaly need to go from SemVer
   (1.2.3) to a tag (v1.2.3)
3. Specific gem dependencies from gitlab-rails need to be added

Prior to this commit, in order to generate CNG variables we would take
the Omnibus map and then add the gem definitions that CNG inspected,
handling item 3 above. However, at this point the CNG map wasn't
representative of what the final product would be because it still
hadn't handled cases 1 and 2. As a result, we were having to avoid the
`MAILROOM_VERSION` entry at multiple points.

Now, we take the Omnibus map and modify it for CNG as soon as the map is
requested via `get_cng_versions`. At that point, it's in the format that
the `cng_version_changes?` and `update_cng` methods expect when they
iterate through it.