Separate Omnibus and CNG tagging in PassingBuild
There is still an issue to be addressed with how we generate the tag
name for the CNG image. It's currently re-appropriating the `Naming.tag`
method, which assumes the second SHA fragment is an Omnibus commit,
which doesn't make sense for CNG.

It also outlines a disparity in that we don't tag the Deployer when we
tag a CNG image.