Fix linking of auto-deploy QA issues
When creating QA issues the code was expecting to find a monthly release
issue. For auto-deploy QA issues there is no release issue that uses the
exact same version, instead they should be linked to the corresponding
MAJOR.MINOR release issue.

We also take care of not linking the parent issue if it does not (yet)
exist.

This fixes https://gitlab.com/gitlab-org/release-tools/issues/388