Simplify method arguments for ComponentVersions class
Previously, this class was working with either GitLab FOSS or GitLab EE.
After the single codebase, it only ever works with EE, so passing around
a `project` argument was unnecessary.

This also moves `GitlabClient.file_contents` to a non-delegate method so
that we can stop manually calling (and stubbing) `project_path`
everywhere we use it.