Notify released merge requests for RCs
Not notifying merge requests for RCs meant that when we create the final
.0 release, there are no merge requests associated with the deployment
(as they are already associated with a previous deployment). This meant
that in practise we would never notify released merge requests.

To fix this, we simply notify merge requests for RC deployments. Since
reverts after the first RC are rare, this should be good enough for 99%
of the cases.

As part of this we also add some logging to make it easier to debug the
code in CI.

This fixes https://gitlab.com/gitlab-org/release-tools/issues/395.