Handle QA pipelines when finding green commits
QA pipelines have a small pipeline, with only 11 jobs (at the time of
writing), and this amount may change over time. Such QA pipelines
_should_ be considered when trying to tag an auto-deployment release, as
issues with our QA tests could affect the deployer.

Our old logic for excluding small pipelines involved checking the number
of jobs, specifically if this number was less than 50. This would result
in QA pipelines being ignored.

To work around this, we now check for the presence of the
"setup-test-env" job. This job is present in both full and QA-only
pipelines, but not in pipelines for documentation-only changes.

We also add some additional logging so it's easier to debug this code in
the future.