Refactor the ComponentVersions update methods
Previously the `PassingBuild` class would ask this class if it had
changes, and then if it did, told it to update, rendering a log message
with the ensuing commit.

Now, we just tell it to update. If it doesn't have changes, it does
nothing. If it has changes, it performs the commit (outside of dry-runs)
and logs the commit URL.