Move Omnibus deployment into Omnibus tagger
If we successfully created a tag, we always wanted to create the same
tag on the Deployer project.

Again, this logic could be different based on the packager, and CNG may
want to deploy differently, so this logic has been isolated
per-packager.

This greatly reduces the complexity of the PassingBuild class.