Verify MRs targeting master close security issue
Adds a validation into MergeRequestValidator to verify if the merge
requests targeting the default project's branch (master) will close
the security implementation issue when merged.

Closing issues automatically only works if the merge requests target the
default branch (master) in this case, so the validation is skipped if
the merge request targets a different branch

This later can be used on
https://gitlab.com/gitlab-org/release-tools/-/issues/369 to verify how
many issues were closed after they were processed.