Limit latest_successful_on_build to the merge-base
This commit introduce an optional limit to
Commits#latest_successful_on_build, if the limit is povided and
merge_base_limit feature flag is set, then the search for a successfull
build will be limited to that provided commit id (excluded).

During an auto-deploy tagging, we'll set such limit as the merge-base
with master.

This will prevent gitaly downgrades