Correct names for tracking auto-deploy releases
For software released as a CE and EE version, we now include "-ee" and
"-ce" in the names used for the JSON release data. For example, Omnibus
data is now under "omnibus-gitlab-ee" instead of just "omnibus-gitlab".

This change is needed so we can start tracking CE releases, without the
names of releases being confusing (e.g. "gitlab" and "gitlab-ce").