Refactor generating of auto-deploy tags
Auto-deploy tags include a timestamp, which was generated using the
creation time of a packager's (e.g. CNG) auto-deploy branch. Relying on
branch creation timings means that packagers can end up with different
tag names, a problem discussed in
https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/794. These
different tag names also prevent us from recording release data in the
same file.

To work around this, we refactor the auto-deploy code such that CNG and
Omnibus use the same timestamp for generating their tags. This is
implemented by passing an instance of AutoDeployBranch around, and
making this instance in charge of generating a timestamp. This also
allows us to remove some duplicate code used for extracting the major
and minor version from an auto-deploy branch name.