Track release data of monthly releases
This adds support for tracking release data of monthly releases, RCs,
and patch releases; in addition to tracking auto-deploys.

To facilitate this, a few changes have been made to the process
overall:

1. Release data is now tracked in a single file for all packagers,
   instead of each packager having its own directory (e.g. releases/cng
   for CNG builds). This makes it easier to track data for monthly
   releases. Release data now lives at "releases/MAJOR/FILE.json".

2. For auto-deploys, the JSON file names use a normalised version. This
   version is in the format "MAJOR.MINOR.TIME", where TIME is the
   timestamp included in the auto-deploy tag names. The Omnibus and
   GitLab SHAs are not included in the file names.

3. For monthly releases, the JSON file names use the normalised patch
   version, so "MAJOR.MINOR.PATCH.json" or "MAJOR.MINOR.PATCH-rcX.json"
   for RCs.

4. Release data is now uploaded in the release.rake and auto_deploy.rake
   Rake files; instead of being uploaded in the release code itself.

The changes needed to make this work are not the most pretty of changes,
but it's the best we can do with our current setup. A plan on how to
deal with this can be found in the GitLab comment
https://gitlab.com/groups/gitlab-com/gl-infra/-/epics/113#note_329004701.