Execute security:validator in an independent way
Starting with the current security release, we're now validating
security merge requests in batches, meaning we validate security merge
requests associated with a security issue. These changes are under a
feature flag, and they're preventing our 'security:validate' cron task
to be executed correctly.

To fix the above, we now introduce a new class (ProjectsValidator), that
will be in charged to validate security merge requests associated to a
project. This is the task that will be executed on 'security:validate'

Related to https://gitlab.com/gitlab-org/release-tools/-/issues/426