Handle release dates after the 22nd
When obtaining a version for a release date in
ReleaseManagers::Schedule#version_for_date, we now handle dates that are
after the 22nd of a month. Prior to these changes, requesting the
version for (for example) the 23rd would produce a version of that
month, instead of producing a version for the _next_ month. This commit
fixes this problem, effectively treating the 22nd as the last day of a
month.

This fixes https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/821