Switch to #f_upcoming_release slack channel
This commit removes the usage of monthly release channels
(i.e. f_release_13_0) in favor of a single channel #f_upcoming_release

Historically we used the channel separation in order the let the old
release managers take care of patch release and let the new release
manager take care of the ongoing monthly release.

This is no longer the case, release manager on duty takes care of all
the releases, and this separation only adds confusion and monthly
chores.

It's not uncommon in the first week of the release cycle to have
messages spread across the two channels.