Fixes cherry-picking failure on security service
Previously we used the merge requests associated to an issue, but none
of those had the merge_commit_sha information generating an invalid URL
on the POST request.

This commit sends the merge result instead that contains a
merge_commit_sha.

CherryPicker service implementation was also simplified because of it.

Related to https://gitlab.com/gitlab-org/release-tools/-/issues/431
and to https://gitlab.com/gitlab-org/release-tools/-/issues/429